requirements.txt
category-encoders==2.6.3
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
imbalanced-learn==0.12.2
imblearn==0.0
joblib==1.4.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
statsmodels==0.14.1
threadpoolctl==3.4.0
ttkthemes==3.2.2
tzdata==2024.1
xgboost==2.0.3
